implemented CI fix
